allow to pass REST API KEY via env variable
